+++ test/boruta/oauth/integration/implicit_grant_test.exs
@@ -64,8 +64,8 @@ defmodule Boruta.OauthTest.ImplicitGrantTest do
         error: :invalid_client,
         error_description: "Invalid client_id or redirect_uri.",
         status: :unauthorized,
-        format: :fragment,
-        redirect_uri: "http://redirect.uri"
+        format: nil,
+        redirect_uri: nil
       }}
     end
 
@@ -84,8 +84,8 @@ defmodule Boruta.OauthTest.ImplicitGrantTest do
         error: :invalid_client,
         error_description: "Invalid client_id or redirect_uri.",
         status: :unauthorized,
-        format: :fragment,
-        redirect_uri: "http://bad.redirect.uri"
+        format: nil,
+        redirect_uri: nil
       }}
     end